Rename: 컴포넌트 폴더 명 및 파일명 변경
