+++ app/(route)/(auth)/signup/page.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
 
-import SignUpForm from './_components/sign-up-form';
+import SignUpForm from './_components/SignUpForm';
 
 function SignUP() {
   return (
